requirements.txt
geopy==2.4.1
uvicorn==0.20.0
python-multipart==0.0.20
pydantic==2.10.6
fastapi==0.115.11
jinja2==3.1.2
python-dotenv==1.1.0
SQLAlchemy==2.0.40
pydantic-settings==2.8.1
asyncpg==0.30.0
tronpy==0.5.0
starlette==0.46.1
pytest==8.3.5
polyfactory==2.6.3
Faker==19.2.0
factory_boy==3.3.0
pytest-asyncio==0.26.0
httpx==0.28.1
psycopg2-binary==2.9.5
psycopg2==2.9.10